requirements.txt
Django==1.10
djangorestframework==3.7.7
djangorestframework-bulk==0.2.1
